+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Graph: remove excessive warning when scanning all elements.
 * Graph: add methods to get vertices and edges in a range of IDs
 * Accumulo: add methods to get the tablet splits
+* Elasticsearch: Add extended bounds to histograph aggregation
 
 # v2.4.0